+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/api/ConfiguredEqualsVerifier.java
@@ -11,6 +11,7 @@ import nl.jqno.equalsverifier.Func.Func2;
 import nl.jqno.equalsverifier.Warning;
 import nl.jqno.equalsverifier.internal.instantiation.vintage.FactoryCache;
 import nl.jqno.equalsverifier.internal.instantiation.vintage.PrefabValuesApi;
+import nl.jqno.equalsverifier.internal.reflection.PackageScanOptions;
 import nl.jqno.equalsverifier.internal.reflection.PackageScanner;
 import nl.jqno.equalsverifier.internal.util.ListBuilders;
 import nl.jqno.equalsverifier.internal.util.Validations;
@@ -151,11 +152,15 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * instead.
      *
      * @param packageName A package for which each class's {@code equals} should be tested.
+     * @param options     Modifications to the standard package scanning behaviour.
      * @return A fluent API for EqualsVerifier.
      */
     @CheckReturnValue
-    public MultipleTypeEqualsVerifierApi forPackage(String packageName) {
-        return forPackage(packageName, false);
+    public MultipleTypeEqualsVerifierApi forPackage(String packageName, ScanOption... options) {
+        PackageScanOptions opts = ScanOptions.process(options);
+        List<Class<?>> classes = PackageScanner.getClassesIn(packageName, opts);
+        Validations.validatePackageContainsClasses(packageName, classes);
+        return new MultipleTypeEqualsVerifierApi(classes, this);
     }
 
     /**
@@ -168,12 +173,12 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param packageName     A package for which each class's {@code equals} should be tested.
      * @param scanRecursively true to scan all sub-packages
      * @return A fluent API for EqualsVerifier.
+     * @deprecated Use {@link #forPackage(String, ScanOption...)} instead.
      */
     @CheckReturnValue
+    @Deprecated
     public MultipleTypeEqualsVerifierApi forPackage(String packageName, boolean scanRecursively) {
-        List<Class<?>> classes = PackageScanner.getClassesIn(packageName, null, scanRecursively);
-        Validations.validatePackageContainsClasses(packageName, classes);
-        return new MultipleTypeEqualsVerifierApi(classes, this);
+        return scanRecursively ? forPackage(packageName, ScanOption.recursive()) : forPackage(packageName);
     }
 
     /**
@@ -189,10 +194,14 @@ public final class ConfiguredEqualsVerifier implements EqualsVerifierApi<Void> {
      * @param packageName A package for which each class's {@code equals} should be tested.
      * @param mustExtend  if not null, returns only classes that extend or implement this class.
      * @return A fluent API for EqualsVerifier.
+     * @deprecated Use {@link #forPackage(String, ScanOption...)} with {@link ScanOption#mustExtend(Class)}, and
+     *                 possibly {@link ScanOption#recursive()}, instead.
      */
     @CheckReturnValue
+    @Deprecated
     public MultipleTypeEqualsVerifierApi forPackage(String packageName, Class<?> mustExtend) {
-        List<Class<?>> classes = PackageScanner.getClassesIn(packageName, mustExtend, true);
+        PackageScanOptions opts = ScanOptions.process(ScanOption.recursive(), ScanOption.mustExtend(mustExtend));
+        List<Class<?>> classes = PackageScanner.getClassesIn(packageName, opts);
         Validations.validatePackageContainsClasses(packageName, classes);
         return new MultipleTypeEqualsVerifierApi(classes, this);
     }